+++ src/components/Header/UserLogoModal.jsx
@@ -1,23 +1,15 @@
 import css from './UserLogoModal.module.scss';
 import { useState } from 'react';
 import { FiEdit2 } from 'react-icons/fi';
-import { useLogoutMutation } from 'redux/authSlice';
 import { notification } from 'components/Shared/notification';
 import UserInfoModal from './UserInfoModal';
-
+import LogoutBtn from './LogoutBtn';
 
 const UserLogoModal = () => {
-  const [isDisabled, setIsDisabled] = useState(false);
   const [showInfoModal, setShowInfoModal] = useState(false);
-  const [dispatch] = useLogoutMutation();
 
   const handleEditProfileClick = () => {
     setShowInfoModal(!showInfoModal);
-  };
-
-  const handleSignoutClick = () => {
-    setIsDisabled(true);
-    dispatch();
   };
 
   return (
@@ -31,16 +23,11 @@ const UserLogoModal = () => {
           <p className={css.text}>Edit profile</p>
           <FiEdit2 height="14px" width="14px" />
         </button>
-        <button
-          type="button"
-          className={`${css.btn} ${css.logoutBtn}`}
-          onClick={handleSignoutClick}
-          disabled={isDisabled}
-        >
-          Log out
-        </button>
+        <LogoutBtn />
       </div>
-      {showInfoModal && <UserInfoModal toggleModalShown={handleEditProfileClick}/>}
+      {showInfoModal && (
+        <UserInfoModal toggleModalShown={handleEditProfileClick} />
+      )}
     </>
   );
 };
